+++ Changelog.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.0.8] - 2017-03-01
+- One single cura for all dagoma printers.
+- Model color according to the chosen filament.
+
+
 ## [1.0.7] - 2017-12-01
 - Fix crashes when changing a parameter during the slicing process.
 - Add a filament color combo box depending on the chosen filament.